fix(screengrab): use more robust args to test grim
Screenshotting a small region of the screen to test grim with
a minimum delay turned out to be not robust, because a pre-defined
region might not be part of the display.

Therefor we switched to grab the whole screen for testing, but scale
the result to a tiny picture to keep the delay low.